I1BCIP-155: Moved profile page code to user details page and removed profile page
