+++ src/App.tsx
@@ -12,10 +12,10 @@ import ContactDetails from './pages/contacts/ContactDetails';
 import Accounts from './pages/accounts/Accounts';
 import AccountDetails from './pages/accounts/AccountDetails';
 import Users from './pages/users/Users';
-import ProfilePage from './pages/profile/Profile';
 import Deals from './pages/deals/Deals';
 import DealsCardView from './pages/deals/DealsCardView';
 import DealDetails from './pages/deals/DealDetails';
+import UserDetails from './pages/users/UserDetails';
 
 function App() {
   return (
@@ -42,11 +42,11 @@ function App() {
             <Route path="accounts" element={<Accounts />} />
             <Route path="accounts/:accountId" element={<AccountDetails />} />
             <Route path="users" element={<Users />} />
-            <Route path="users/:id" element={<ProfilePage />} />
+            <Route path="users/:id" element={<UserDetails />} />
             <Route path="deals" element={<Deals />} />
             <Route path="deals/card-view" element={<DealsCardView />} />
             <Route path="deals/:dealId" element={<DealDetails />} />
-            <Route path="profile/:id" element={<ProfilePage />} />
+            <Route path="profile/:id" element={<UserDetails />} />
           </Route>
         </Routes>
       </Router>
