+++ src/pages/users/styles.css
@@ -1,8 +1,8 @@
 .profile-page-container{
     width: 100%;
-    height: calc(100vh - 60px);
-    margin-top: 60px;
+    height: 100%;
     padding: 24px;
+    padding-top: 16px;
     box-sizing: border-box;
 }
 
@@ -13,7 +13,7 @@
     align-items: center;
     padding-right: 32px;
     justify-content: flex-end;
-    margin-bottom: 12px;
+    margin-bottom: 16px;
     height: 36px;
 }
